[JDR-Major-Calculate-Action] Change calculate function to dispatch action
