+++ src/store/store.ts
@@ -1,10 +1,10 @@
 import { createStore, applyMiddleware, compose } from 'redux';
 import { createLogger } from 'redux-logger';
+import multi from 'redux-multi';
+import thunk from 'redux-thunk';
 import rootReducers from './root-reducers';
 
-const middlewares = [];
-
-middlewares.push(createLogger());
+const middlewares = [createLogger(), multi, thunk];
 
 const enhancer = compose(applyMiddleware(...middlewares));
